refactor: replace AutoMigrate with MigrateUp/MigrateDown in subscription store migration and update dependencies
Replace AutoMigrate() with MigrateUp() in subscription store migration Up method. Implement Down method with proper store initialization checks and MigrateDown() call instead of no-op. Add cmsstore local replace directive. Bump dependencies including auditstore (v1.5.0), blindindexstore (v1.11.0), blogstore (v1.24.0), cachestore (v1.5.0), chatstore (v1.0.0), customstore (v1.9.0), feed

+++ pkg/logadmin/log_manager/log_list_results.go
@@ -25,7 +25,7 @@ type logListFilters struct {
 // logListResult represents the result of a log listing.
 type logListResult struct {
 	Logs    []logstore.LogInterface
-	Total   int
+	Total   int64
 	HasMore bool
 }
 
@@ -87,7 +87,7 @@ func listLogs(app app.AppInterface, f logListFilters) (logListResult, error) {
 	}
 
 	// Total count for pagination
-	total := 0
+	total := int64(0)
 	if n, err := app.GetLogStore().LogCount(context.Background(), query); err == nil {
 		total = n
 	}
